Ordering, Match and cmp
This commit introduces the match expression. Here we are able to pass
the arms located inside the {} to the match function. the match function
takes the ouput of .cmp's evaluation of guess and secret_number and
determines which println to run. This is very much like a java switch
statement.

Note: this code will not compile and run as secret_number and guess are
two differet types. The next commit will solve this problem.